Allows default per kw var
I made a mistake when I implemented the keyword vars. When the user uses
"#$keyword3|def1## #$keyword3|def2##" only the first default is picked
up and used as a default for keyword3. The intention was for each
instance of a missign keyword to be able to have its own default. The
reason being that depending on where you sprinkle that keyword in the
query structure, it may impact how the Lucene query is parsed if a
keyword is missing. IE the Solr default of "" would end up screwing up a
phrase query. However something like "" would be an ok default in a
Lucene query.