Adds syntax for computed fields
In Solr, its convenient to break apart function queries into pieces in a
syntax like

    partialFunc=product(tf('text','law'),2.0)

but what partialFunc is gets lost in the explanation.

A simple solution to this is to allow this function to come back as a
computed field. So we've introduced the fieldSpec syntax

    function:partialFunc

which is intended to display the result of partialFunc in
Splainer/Quepid's output.