normalDocSvc consumes function values as "subs"
To the consumer of this API, this document is just something to display.
So the results of functions, for now, are just another kind of subField
to display.

We need to do a little bit of work as we send to Solr "foo:$foo" but
Solr just returns back "foo". So we need to ensure that we display as
"foo"