Ignores spurious function_score query
Elasticsearch's function_score queries layer on all kinds of cruft that
are typically special cases. For example the "max_boost" is an argument
taht can put a ceiling on your boost. Also you always have a
"queryBoost" of 1, even if you didn't specify any kind of boost.

Experimenting with o19s's blog, this seems to give the sanest results
for function_score queries