Allows URI decoding to fail gracefully
If you happen to put mm=50% in a Solr query in splainer, splainer just
gets stuck. This is because URI decoding throws an excpetion. the
decodeURIcomponent can tell this isn't a legit decoded URI component.
Unfortunately for Solr users, its easy to have a giant query URL where
mm=50% is buried in at character 500. Its a royal pain to find that mm
and manually escape it. Moreover, we advertise you can just "paste a URL
into splainer" and it'll work.

All that to day, we should be more forgiving in this error case.