Smarter handling of Basic Auth and encoded characters. (#152)
Smarter handling of basic auth, especially when you have special characters like $ and @ in the password.
For all api methods except JSONP, we use the Authorization header, even if you pass in a url with the embedded basic auth like http://user:pass@myserver.com.
Then for JSONP, which does NOT support headers, we embed the basic auth back into the URL.
FYI: Chrome does NOT allow you to use embedded username password in the URL.